+++ src/ehrdata/core/constants.py
@@ -1,1 +1,1 @@
-R_LAYER_KEY = "r_layer"
+R_LAYER_KEY = "R_layer"