+++ src/ehrdata/dt/datasets.py
@@ -299,7 +299,7 @@ def physionet2012(
     obs.index = obs.index.astype(str)
     var.index = var.index.astype(str)
 
-    edata = EHRData(r=r, obs=obs, var=var, t=t)
+    edata = EHRData(R=r, obs=obs, var=var, t=t)
 
     return edata[~edata.obs.index.isin(drop_samples or [])]