+++ tests/conftest.py
@@ -1,30 +1,42 @@
 import anndata as ad
+import dask.array as da
 import duckdb
 import numpy as np
 import pandas as pd
 import pytest
+import sparse as sp
 
 from ehrdata import EHRData
 from ehrdata.io.omop import setup_connection
 
 
 @pytest.fixture
-def X_32():
+def X_numpy_32():
     return np.arange(1, 7).reshape(3, 2)
 
 
 @pytest.fixture
-def X_33():
+def X_numpy_33():
     return np.arange(1, 10).reshape(3, 3)
 
 
 @pytest.fixture
-def r_322():
+def R_numpy_322():
     return np.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]]])
 
 
 @pytest.fixture
-def r_333():
+def R_sparse_322():
+    return sp.COO(np.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]]]))
+
+
+@pytest.fixture
+def R_dask_322():
+    return da.from_array(np.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]]]), chunks=(3, 2, 2))
+
+
+@pytest.fixture
+def R_numpy_333():
     return np.arange(1, 28).reshape(3, 3, 3)
 
 
@@ -59,13 +71,13 @@ def t_31():
 
 
 @pytest.fixture
-def edata_333(X_33, r_333, obs_31, var_31, t_31):
-    return EHRData(X=X_33, r=r_333, obs=obs_31, var=var_31, t=t_31)
+def edata_333(X_numpy_33, R_numpy_333, obs_31, var_31, t_31):
+    return EHRData(X=X_numpy_33, R=R_numpy_333, obs=obs_31, var=var_31, t=t_31)
 
 
 @pytest.fixture
-def adata_33(X_33, obs_31, var_31):
-    return ad.AnnData(X=X_33, obs=obs_31, var=var_31)
+def adata_33(X_numpy_33, obs_31, var_31):
+    return ad.AnnData(X=X_numpy_33, obs=obs_31, var=var_31)
 
 
 @pytest.fixture